Beginning of update to GiantPipe 2.0
The new structure of GiantPipe is build using a mix of Object-Oriented and Functional design principles. It it intended to be more readable and user-friendly, as well as allowing future code expansion and maintenance.

This first commit is focusing on the RegisterMaps() function from version 1. It will be useful to update this first so that there is a robust way of operating on many VISIR datasets at once. This functionality will be finished later and the structure of classes and functions in the filesystem will be decided then as well.