Built the ProcessData class for reading FITS files, and constructing data and metadata for future processing modules.
Re-created the functionality of the old RegisterMaps() under the hybrid Object-Oriented-Functional (OOF) scheme.

The code is now able to read FITS files and construct data maps and metadata from them.

This new code is MUCH faster than the old one (without benchmarking). The old structure was inherited from the previous IDL code, but even that code didn't actually need to be so slow).

This code is intended to be used to interface with all data in an agnostic way but is currently coupled quite heavily to the overall structure of the VISIR images and DRM outputs. However, it is built to be expandable for future datasets.