Continued development of register_maps(), begun work on distributing modules.
My Macbook is not so healthy, so this commit should be considered more as a back up, for insurance purposes.

The methods were tweaked to specific purposes (instance, class, and static).

Began breaking up modules from main() in giantpipe.py. Moved to-date work to the register_maps() function in the processing module. Added a testing module with Clock and Profiler classes, useful for benchmarking the code. Separated the Config class to live in the config module.