Added central meridian binning, reading from NetCDF
Could not get the xarray.Dataset.variable.sel() functionality to work, so I coded my own degrees to indices conversion function. It is fine and fast but perhaps look at this in future. Used np.mean(array, axis=1) to do a vectorised zonal mean and avoid using a for loop.