Added 7.9 micron doppler correction
We perform a 1D interpolation at processing.Process.doppler_correction_of_radiance.interpolate_doppler_profile(). We did some quick testing of a module-level class containing the doppler correction in an attempt to cache the calculation and resulting interpolation object but it reduced readability** at the benefit of not much increase in computation speed. So we are going with an in-line computation for each 7-micron image.

**and honestly I couldn't get it working and the debugging was costing me more time than I would like (consider in future iterations if needed)